fix: update domain from vercel.app to json-animation-viewer.com for SEO
- Replace all vercel.app URLs with json-animation-viewer.com in:
  - layout.tsx (metadataBase, OG url, google verification code)
  - robots.ts (sitemap URL)
  - sitemap.ts (page URL)
  - JsonLd.tsx (schema.org screenshot URL)
- Add canonical redirect from vercel.app to custom domain in next.config.ts
- Refactor page.tsx to server component by extracting interactive parts
  into AnimationViewer client component (improves SSR/indexability)
- Fix google verification code placeholder to actual value
