+++ src/app/robots.ts
@@ -6,6 +6,6 @@ export default function robots(): MetadataRoute.Robots {
       userAgent: "*",
       allow: "/",
     },
-    sitemap: "https://json-animation-viewer.vercel.app/sitemap.xml",
+    sitemap: "https://json-animation-viewer.com/sitemap.xml",
   };
 }
