+++ src/app/sitemap.ts
@@ -3,7 +3,7 @@ import { MetadataRoute } from "next";
 export default function sitemap(): MetadataRoute.Sitemap {
   return [
     {
-      url: "https://json-animation-viewer.vercel.app",
+      url: "https://json-animation-viewer.com",
       lastModified: new Date(),
       changeFrequency: "weekly",
       priority: 1,
